backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.35
alembic==1.12.1
pydantic>=2.5.1
pydantic-settings>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.1
apscheduler==3.10.4
python-multipart==0.0.6
deep-translator>=1.11.0
feedparser>=6.0.0
